==> 8_runner.py <==
import os

# os.system('python 0_clear_dir.py')

# os.system('python 1_collection.py')

# os.system('python 1_collection_test.py')

# os.system('python 1_file_checker.py')

# os.system('python 1A_tweet_data_collection_process.py')

# os.system('python 1B_user_checker.py')

# os.system('python 1C_follower_graph_process.py')

# os.system('python 2_distributions.py')

# os.system('python 3A_with_fg.py')

# os.system('python 3A_without_fg.py')

# os.system('python 4_cleaning.py')

# os.system('python 5_tuning.py')

# os.system('python 6_tuning.py')

# os.system('python 7_results.py')
